Made code leaner by outsourcing all logic from files in routes/api folder to controllers in the controllers folder. This gives the app a MVC like architecture and the makes the code leaner
